+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Show linebreaks in quotes
 - trust all labeled links from device chat
 - Adjust order of buttons at labeled link dialog
+- call `maybeNetwork` when waking up from standby (macOS and Windows)
 
 ## Fixed
 - Fix "copy link" context menu option for labeled links